deleted class CreateNewViewTest and moved testMoveFolderToNewViewList to existed NewViewTest
